# docs/swagger.yaml
basePath: /api
definitions:
  api.JobStatus:
    enum:
    - queued
    - processing
    - completed
    - failed
    type: string
    x-enum-varnames:
    - JobStatusQueued
    - JobStatusProcessing
    - JobStatusCompleted
    - JobStatusFailed
  api.VideoJob:
    properties:
      completed_at:
        type: string
      created_at:
        type: string
      error:
        type: string
      failed_batches:
        type: integer
      filename:
        type: string
      full_text:
        description: Combined transcription
        type: string
      id:
        type: string
      processed_batches:
        type: integer
      status:
        $ref: '#/definitions/api.JobStatus'
      successful_batches:
        type: integer
      total_batches:
        type: integer
      total_frames:
        type: integer
      transcription:
        description: Accumulated text results
        items:
          type: string
        type: array
      updated_at:
        type: string
      video_info:
        additionalProperties: true
        type: object
    type: object
  api.WebSocketMessage:
    properties:
      confidence:
        type: number
      draft_text:
        type: string
      enhanced:
        type: boolean
      final_text:
        type: string
      first_sequence:
        type: integer
      full_text:
        type: string
      last_sequence:
        type: integer
      literal_text:
        type: string
      segment_id:
        type: integer
      sequence:
        type: integer
      status:
        type: string
      text:
        type: string
      token_count:
        type: integer
      truncated:
        type: boolean
      type:
        type: string
    type: object
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API for video frame streaming and processing via WebSocket and video
    upload
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Video Streaming API
  version: "1.0"
paths:
  /health:
    get:
      description: Check if the API is running and healthy
      produces:
      - text/plain
      responses:
        "200":
          description: OK
          schema:
            type: string
        "503":
          description: Server is draining
          schema:
            type: string
      summary: Health check endpoint
      tags:
      - health
  /job/{id}:
    get:
      description: |-
        Poll this endpoint to check the status of a video processing job and retrieve results when complete

        **Job Status Values:**
        - `queued`: Job created, waiting to start
        - `processing`: Currently extracting and processing frames
        - `completed`: All processing done, transcription available
        - `failed`: Processing encountered an error

        **Response includes:**
        - Current status and progress (processed_batches / total_batches)
        - Accumulated transcription results
        - Full combined text when completed
        - Video metadata and statistics
      parameters:
      - description: Job ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Job status and results
          schema:
            $ref: '#/definitions/api.VideoJob'
        "404":
          description: Job not found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get video processing job status
      tags:
      - video
  /socket:
    get:
      consumes:
      - application/octet-stream
      description: |-
        Establishes a WebSocket connection for receiving video frames. Send binary frames to the server, and receive text responses back.

        **Client Flow:**
        1. Connect to same-origin `/api/socket` using `wss:` on HTTPS pages and `ws:` otherwise
        2. Send video frames as binary messages (MessageBinary)
        3. Server buffers frames and sends batches of 32 to processing API
        4. Receive ordered gesture, formatting, and transcript events as JSON text messages

        **Two-layer transcript:**
        Every stable gesture is emitted immediately as type=gesture with status=draft. Stable gestures are grouped after 3 seconds of idle time or 6 tokens. A type=formatting event marks a segment being processed. The final type=transcript event replaces that segment's raw draft with enhanced or literal text. OpenRouter never blocks newer gesture events.

        **Gesture event:**
        ```json
        {
        "type": "gesture", "status": "draft", "text": "работать",
        "final_text": "", "draft_text": "я работать", "full_text": "я работать", "literal_text": "я работать",
        "confidence": 0.91, "sequence": 2, "segment_id": 1
        }
        ```

        **Final segment event:**
        ```json
        {
        "type": "transcript", "status": "enhanced", "enhanced": true,
        "text": "Я работаю.", "final_text": "Я работаю.", "draft_text": "",
        "full_text": "Я работаю.", "literal_text": "я работать", "sequence": 2, "segment_id": 1,
        "first_sequence": 1, "last_sequence": 2, "token_count": 2,
        "confidence": 0.91
        }
        ```
        full_text is the authoritative rendered snapshot and equals finalized presentation text followed by all raw draft tokens. literal_text independently preserves recognizer-only output. text and confidence remain for compatibility. Sequences and segment IDs are monotonic per connection. truncated=true means the bounded session discarded its oldest finalized prefix.

        **Frontend Example:**
        ```javascript
        const wsURL = new URL('/api/socket', window.location.origin);
        wsURL.protocol = window.location.protocol === 'https:' ? 'wss:' : 'ws:';
        const ws = new WebSocket(wsURL);

        // Send binary frame data
        ws.send(frameDataBlob);

        // Receive text messages
        ws.onmessage = (event) => {
        const data = JSON.parse(event.data);
        console.log('Received text:', data.text);
        };
        ```
      produces:
      - application/json
      responses:
        "101":
          description: Ordered two-layer transcript event
          schema:
            $ref: '#/definitions/api.WebSocketMessage'
        "400":
          description: Bad Request - Failed to accept websocket
          schema:
            type: string
        "503":
          description: WebSocket capacity exhausted or server draining
          schema:
            type: string
      summary: WebSocket endpoint for video frame streaming
      tags:
      - websocket
  /upload:
    post:
      consumes:
      - multipart/form-data
      description: |-
        Upload a video file and receive a job ID immediately. The video will be processed asynchronously in batches of 32 frames.

        **Process Flow:**
        1. Upload video file via multipart form data
        2. Receive job ID immediately (status: queued)
        3. Server processes video in background
        4. Poll /job/{id} endpoint to check status and get results
        5. When complete, full transcription is available

        **Immediate Response Format:**
        ```json
        {
        "job_id": "550e8400-e29b-41d4-a716-446655440000",
        "status": "queued",
        "message": "Video upload accepted, processing started"
        }
        ```

        **Frontend Example:**
        ```javascript
        // 1. Upload video
        const formData = new FormData();
        formData.append('video', videoFile);
        const apiBase = '/api';
        const uploadResp = await fetch(`${apiBase}/upload`, {
        method: 'POST',
        body: formData
        });
        const { job_id } = await uploadResp.json();

        // 2. Poll for results
        const pollInterval = setInterval(async () => {
        const statusResp = await fetch(`${apiBase}/job/${job_id}`);
        const job = await statusResp.json();

        if (job.status === 'completed') {
        console.log('Transcription:', job.full_text);
        clearInterval(pollInterval);
        } else if (job.status === 'failed') {
        console.error('Processing failed:', job.error);
        clearInterval(pollInterval);
        }
        }, 2000);
        ```
      parameters:
      - description: Video file to process
        in: formData
        name: video
        required: true
        type: file
      - description: 'Minimum frame extraction interval (default: 1; automatically
          increased to keep extraction bounded)'
        in: formData
        name: interval
        type: integer
      produces:
      - application/json
      responses:
        "202":
          description: Job created and processing started
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request - Invalid file or format
          schema:
            additionalProperties:
              type: string
            type: object
        "413":
          description: Video exceeds 100 MiB
          schema:
            additionalProperties:
              type: string
            type: object
        "415":
          description: Unsupported or invalid video
          schema:
            additionalProperties:
              type: string
            type: object
        "429":
          description: Video processor is busy
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
        "503":
          description: Upload capacity exhausted or server draining
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Upload video for async frame-by-frame processing
      tags:
      - video
schemes:
- http
- https
- ws
- wss
swagger: "2.0"